# openapi.yaml
openapi: 3.0.3
info:
  title: Travenion API
  description: 旅游计划管理系统API文档
  version: 2.0.0
  contact:
    name: Travenion Team
    email: support@travenion.com
servers:
  - url: /travenion/api
    description: 生产环境
paths:
...
  /plans/{id}/days/{dayId}/attractions/{attractionId}:
     put:
       summary: 更新景点信息
       description: 更新指定景点的信息
       tags:
         - 景点管理
       security:
         - bearerAuth: []
       parameters:
         - in: path
           name: id
           required: true
           schema:
             type: integer
           description: 计划ID
           example: 1
         - in: path
           name: dayId
           required: true
           schema:
             type: integer
           description: 行程日ID
           example: 1
         - in: path
           name: attractionId
           required: true
           schema:
             type: integer
           description: 景点ID
           example: 1
       requestBody:
         required: true
         content:
           application/json:
             schema:
               type: object
               properties:
                 name:
                   type: string
                   description: 景点名称
                   example: "清水寺"
                 address:
                   type: string
                   description: 景点地址
                   example: "京都市东山区清水1丁目294"
                 description:
                   type: string
                   description: 景点描述
                   example: "京都著名的佛教寺庙"
                 visitOrder:
                   type: integer
                   description: 访问顺序
                   example: 1
       responses:
         '200':
           description: 景点更新成功
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/DayAttraction'
         '401':
           description: 未授权
         '404':
           description: 景点未找到
         '500':
           description: 更新失败
     delete:
       summary: 删除景点
       description: 删除指定的景点
       tags:
         - 景点管理
       security:
         - bearerAuth: []
       parameters:
         - in: path
           name: id
           required: true
           schema:
             type: integer
           description: 计划ID
           example: 1
         - in: path
           name: dayId
           required: true
           schema:
             type: integer
           description: 行程日ID
           example: 1
         - in: path
           name: attractionId
           required: true
           schema:
             type: integer
           description: 景点ID
           example: 1
       responses:
         '200':
           description: 景点删除成功
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   message:
                     type: string
                     example: "景点删除成功"
         '401':
           description: 未授权
         '404':
           description: 景点未找到
         '500':
           description: 删除失败
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DayAttraction:
       type: object
       properties:
         id:
           type: integer
           description: 景点记录ID
           example: 1
         dayId:
           type: integer
           description: 所属行程日ID
           example: 1
         name:
           type: string
           description: 景点名称
           example: "清水寺"
         address:
           type: string
           description: 景点地址
           example: "京都市东山区清水1丁目294"
         description:
           type: string
           description: 景点描述
           example: "京都著名的佛教寺庙，以其木制舞台而闻名"
         latitude:
           type: number
           format: float
           description: 纬度
           example: 34.9949
         longitude:
           type: number
           format: float
           description: 经度
           example: 135.7851
         visitOrder:
           type: integer
           description: 访问顺序
           example: 1
         createdAt:
           type: string
           format: date-time
           description: 创建时间
           example: "2024-01-01T00:00:00Z"
         updatedAt:
           type: string
           format: date-time
           description: 更新时间
           example: "2024-01-01T00:00:00Z"